Work on displaying the correct costs in the storeboard and not being able to purchase things that you cannot afford.

+++ Space/Assets/Scripts/Gameplay/Weapon/Pew Pewers/WeaponScript.cs
@@ -68,7 +68,7 @@ public abstract class WeaponScript : MonoBehaviour
 	{
 		return (float)System.Math.Round( statValue, sigFigs );
 	}
-	protected void SetCost(int c)
+	protected void SetCost(float c)
 	{
 		cost = c;
 	}
